requirements.txt
yaml
redis
requests
beautifulsoup4
tqdm
python-dotenv